+++ include/wx/univ/radiobox.h
@@ -88,7 +88,7 @@ public:
     virtual void SetSelection(int n);
     virtual int GetSelection() const;
 
-    virtual int GetCount() const { return m_buttons.GetCount(); }
+    virtual int GetCount() const { return (int) m_buttons.GetCount(); }
     virtual int GetColumnCount() const { return m_numCols; }
     virtual int GetRowCount() const { return m_numRows; }
